+++ src/thread/thread.Message.ts
